requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configobj==5.0.6
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
google-api-core==1.26.2
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
systematic==4.8.7
typing-extensions==3.7.4.3
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
